add .gitignore file to the project
I'm adding this .gitignore file to our project, since Android studio generates some Local files that only use for the Local user and they are not necessary the same or needed for other team menbers, and this file will filter those unneeded files.

So please, verify if I missing some files that should not be uploaded to our repo and add them to the list. After we merge this file, please pull this file before you commit anything.